+++ frontend/src/metabase/dashboard/constants.ts
@@ -3,6 +3,8 @@ import type {
   DashboardState,
 } from "metabase-types/store";
 
+import type { EmbedDisplayParams } from "./types";
+
 export const SIDEBAR_NAME: Record<DashboardSidebarName, DashboardSidebarName> =
   {
     addQuestion: "addQuestion",
@@ -46,3 +48,13 @@ export const DASHBOARD_SLOW_TIMEOUT = 15 * 1000;
 
 export const DASHBOARD_PDF_EXPORT_ROOT_ID =
   "Dashboard-Parameters-And-Cards-Container";
+
+export const DEFAULT_EMBED_DISPLAY_OPTIONS: EmbedDisplayParams = {
+  bordered: false,
+  titled: true,
+  cardTitled: true,
+  hideDownloadButton: null,
+  hideParameters: null,
+  font: null,
+  theme: "light",
+};
