+++ frontend/src/metabase/dashboard/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from "./use-click-behavior-data";
 export * from "./use-dashboard-fullscreen";
-export * from "./use-embed-display-options";
 export * from "./use-dashboard-nav";
 export * from "./use-dashboard-refresh-period";
 export * from "./use-embed-theme";
