--- frontend/src/metabase/dashboard/hooks/use-embed-display-options.ts
@@ -1,11 +1,0 @@
-import type { EmbedDisplayParams } from "../types";
-
-export const DEFAULT_EMBED_DISPLAY_OPTIONS: EmbedDisplayParams = {
-  bordered: false,
-  titled: true,
-  cardTitled: true,
-  hideDownloadButton: null,
-  hideParameters: null,
-  font: null,
-  theme: "light",
-};
